[ALPS07231990] cmdq: fix coverity
1. Uninitialized scalar variable
2. Dereference before/after null check
3. Logically dead code
4. Enum compared against 0
5. Unintentional integer overflow
6. CERT-C Expression

MTK-Commit-Id: 39949008f25286627eca9e654e4d79190ed9e398

CR-Id: ALPS07231990
Feature: cmdq_gce

+++ drivers/misc/mediatek/cmdq/mailbox/mtk-cmdq-mailbox-ext.c
@@ -2631,7 +2631,7 @@ phys_addr_t cmdq_mbox_get_spr_pa(void *chan, u8 spr)
 	struct cmdq_thread *thread =
 		(struct cmdq_thread *)((struct mbox_chan *)chan)->con_priv;
 
-	return thread->gce_pa + CMDQ_THR_BASE + CMDQ_THR_SIZE * thread->idx +
+	return thread->gce_pa + CMDQ_THR_BASE + CMDQ_THR_SIZE * ((u64)thread->idx) +
 		CMDQ_THR_SPR + 4 * spr;
 }
 EXPORT_SYMBOL(cmdq_mbox_get_spr_pa);
